+++ Basil/include/OrthographicCameraController.h
@@ -1,0 +1,34 @@
+/*
+ * This header declares a class that is a wrapper around OrthographicCamera.
+ */
+#pragma once
+
+#include "Events/ApplicationEvent.h"
+#include "Events/MouseEvent.h"
+#include "Renderer/OrthographicCamera.h"
+#include "Timestep.h"
+
+namespace Basil
+{
+	class OrthographicCameraController
+	{
+		public:
+			OrthographicCameraController(float aspectRatio, bool rotation = false);
+			void onUpdate(Timestep timeStep);
+			void onEvent(Event& e);
+			OrthographicCamera& getCamera();
+			const OrthographicCamera& getCamera() const;
+		private:
+			bool onMouseScrolled(MouseScrolledEvent& e);
+			bool onWindowResize(WindowResizeEvent& e);
+			float aspectRatio;
+			float zoomLevel = 1.0f;
+			OrthographicCamera camera;
+			bool rotation;
+			glm::vec3 cameraPosition = { 0.0f, 0.0f, 0.0f };
+			float cameraRotation;
+			float cameraXSpeed;
+			float cameraYSpeed;
+			float cameraRotateSpeed;
+	};
+}